fix : modify my favorite list by change redux setting

+++ Client/src/Modals/ModalWithDrawal.js
@@ -1,9 +1,14 @@
 import axios from 'axios';
 import * as React from 'react';
 import { useState } from 'react';
+import { useNavigate } from 'react-router';
+import { useDispatch } from 'react-redux';
 import './../Styles/Modal.css';
+import { getMyFavoriteMarkers } from './../Actions/index.js';
 
-function ModalBox ({ handleDeleteInfoBtn }) {
+function ModalBox ({ handleDeleteInfoBtn, handleUserInfo }) {
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
   const [withdrawalReason, setWithdrawalReason] = useState('');
   const [password, setPassword] = useState(null);
   const handleWithdrawalReason = (event) => {
@@ -15,22 +20,36 @@ function ModalBox ({ handleDeleteInfoBtn }) {
     setPassword(password);
   };
   const deleteUserInfo = () => {
-    axios({
-      url: 'https://api.open-spot.tk/users',
-      method: 'delete',
-      data: {
-        quitReason: withdrawalReason,
-        password: password
-      },
-      withCredentials: true
-    }).then((res) => {
-      console.log(res);
-      alert('그동안 이용해주셔서 감사합니다 :)');
-      handleDeleteInfoBtn(true);
-    }).catch((err) => {
-      console.log(err);
-      alert('비밀번호가 일치하지 않습니다. 다시 시도해주세요.');
-    });
+    if (withdrawalReason === '') alert('탈퇴 사유를 입력해주세요 !');
+    if (password === '') alert('비밀번호 확인이 필요한 작업이니, 비밀번호를 입력해주세요');
+    else {
+      axios({
+        url: 'https://api.open-spot.tk/users',
+        method: 'delete',
+        data: {
+          quitReason: withdrawalReason,
+          password: password
+        },
+        withCredentials: true
+      }).then((res) => {
+        console.log(res);
+        alert('그동안 이용해주셔서 감사합니다 :)');
+        handleDeleteInfoBtn(false);
+        handleUserInfo({
+          isLogin: false,
+          role: null,
+          name: null,
+          email: null,
+          profile: 'https://api.open-spot.tk/profile.png',
+          oauthLogin: null
+        });
+        dispatch(getMyFavoriteMarkers([]));
+        navigate('/');
+      }).catch((err) => {
+        console.log(err);
+        alert('비밀번호가 일치하지 않습니다. 다시 시도해주세요.');
+      });
+    }
   };
   return (
     <div className="modal-background">
